+++ v1/HitObjects/SpinnerObject.cs
@@ -6,7 +6,7 @@ namespace BMAPI.v1.HitObjects
         public SpinnerObject() { }
         public SpinnerObject(CircleObject baseInstance) : base(baseInstance) { }
 
-        public int EndTime { get; set; }
+        public float EndTime { get; set; }
 
         public override bool ContainsPoint(Point2 Point)
         {
